Split bluetooth models into base_scanner and wrappers
